Separate RemoteInputDestinations from SdkHarnessClient
Make it a member of the `data` package, as it concisely specifies an
input to the Data Service.
